+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Changed
 - Merge PR #157, adding licence attribute to package.json
+- Merge PR #148, removing cloudformation-js-yaml-schema in favour of custom handling of intrinsic functions
+
+### Added
+- Merge PR #148, adding Fn::split functionality
 
 
 ## [1.6.2] - 2018-04-26